requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
django-environ==0.12.0
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0
psycopg2-binary==2.9.10
djangorestframework-gis==1.2.0

